requirements.txt
# Dipendenze per Business Plan Builder

# Framework principale
streamlit>=1.28.0

# Framework LangGraph e LangChain
langgraph>=0.0.15
langchain>=0.1.0
langchain_openai>=0.0.5

# Gestione dei documenti
python-docx>=0.8.11
pypdf>=0.0.2

# Database vettoriale
chromadb>=0.4.18

# Embedding e tokenizzazione
openai>=1.3.0
tiktoken>=0.5.1

# Ricerca e API esterne
requests>=2.31.0
perplexityai>=0.2.0

# Visualizzazione e analisi dati
matplotlib>=3.7.2
pandas>=2.0.3
tabulate>=0.9.0
plotly>=5.18.0
numpy>=1.24.0

# Utilità
python-dotenv>=1.0.0

# Gemini API
google-generativeai>=0.3.0
